feat: sidedrawer - api update to support sidedrawer.trigger
