+++ lib/src/components/sidedrawer/SidedrawerTrigger.tsx
@@ -1,0 +1,4 @@
+import { Trigger } from '@radix-ui/react-dialog'
+import { styled } from '~/stitches'
+
+export const SidedrawerTrigger = styled(Trigger, {})
